Project from Automate the Boring Stuff
In Chapter 8, the author helps you create a random quiz generator. I just tried to write the program out on my own, instead using a dictionary of questions as keys & YES or NO answers as values.

Trying to internalize the code required to make this program work. Some things I'm using that's new for me:

- random module, particularly random.shuffle() and random.sample()
- %s insert inside strings (what's the name for that?)
- ljust(), center(), and rjust()
-open(), write(), close()